pyproject.toml
[project]
name = "distributed-tracing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-exporter-jaeger>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-requests>=0.50b0",
    "opentelemetry-instrumentation>=0.50b0",
    "opentelemetry-sdk>=1.29.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
